+++ views.php
@@ -8,6 +8,7 @@ return [
     'default' => [
         'calendar_ui/icon/' => __DIR__ . '/graphics',
         'moment.js' => __DIR__ . '/vendor/moment/moment/min/moment.min.js',
+        'moment_timezone.js' => __DIR__ . '/vendors/moment/moment-timezone-with-data.js',
         'cui_datetimepicker_js.js' => __DIR__ . '/vendors/datetimepicker/jquery.datetimepicker.full.min.js',
         'cui_datetimepicker_css.css' => __DIR__ . '/vendors/datetimepicker/jquery.datetimepicker.min.css',
         'cui_locale_all.js' => __DIR__ . '/vendors/fullcalendar/locale-all.js',